requirements.txt
python-telegram-bot==21.0.1
python-dotenv>=1.0.0
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0

